+++ app/(app)/dashboard/page.tsx
@@ -1,23 +1,15 @@
+"use client";
+
 import { BsChevronRight } from "react-icons/bs";
 import { FaDirections, FaPlay } from "react-icons/fa";
 import { MdCall } from "react-icons/md";
 import Link from "next/link";
-import Head from "next/head";
-import DashboardLayout from "~layouts/DashboardLayout";
 
 import { Actions, Button, Card, Heading, List, Text } from "~honeycomb";
 
 export default function Page() {
   return (
-    <DashboardLayout>
-      <Head>
-        <title>Dashboard</title>
-        <meta
-          name="description"
-          content="Dashing through the snow"
-          key="desc"
-        />
-      </Head>
+    <>
       <div className="flex flex-col gap-24">
         <section>
           <Card variant="highlight">
@@ -258,6 +250,6 @@ export default function Page() {
           </List>
         </section>
       </div>
-    </DashboardLayout>
+    </>
   );
 }
