+++ app/(marketing)/terms-and-conditions/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return "Hello, maybe use MD to generate web content";
+}
