+++ app/(workflows)/new-session/invite/invite.module.css
@@ -1,13 +1,3 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: auto;
-  height: 100vh;
-  max-width: 40rem;
-  letter-spacing: 1px;
-}
-
 .invitesContainer {
   margin-top: 1rem;
   width: 100%;
