+++ app/(workflows)/workflow-layout.module.css
@@ -1,0 +1,10 @@
+@layer layout {
+  .container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    margin: auto;
+    height: 100vh;
+    max-width: 40rem;
+  }
+}
