+++ lib/apollo.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { ApolloClient, InMemoryCache } from "@apollo/client";
 
 const client = new ApolloClient({
