--- packages/honeycomb/GlobalLayout/index.ts
@@ -1,1 +1,0 @@
-export * from "./GlobalLayout";
