+++ packages/honeycomb/index.ts
@@ -1,8 +1,10 @@
+"use client";
+
 export * from "./Actions";
 export * from "./Button";
 export * from "./Card";
 export * from "./Container";
-export * from "./GlobalLayout";
+export * from "./PageLayout";
 export * from "./Heading";
 export * from "./List";
 export * from "./Logo";
