--- pages/login/index.tsx
@@ -1,9 +1,0 @@
-import Link from "next/link";
-
-export default function Page() {
-  return (
-    <>
-      <h1>Hello, Login Page</h1> <Link href="/dashboard">Login</Link>
-    </>
-  );
-}
