+++ plop-templates/basic-component/{{pascalCase componentName}}.tsx.hbs
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { HTMLAttributes } from "react";
 import clsx from "clsx"
