+++ styles/global.css
@@ -1,4 +1,4 @@
-@layer base, components, utilities;
+@layer base, layout, components, utilities;
 
 @import url("https://fonts.googleapis.com/css2?family=DM+Sans:wght@400;700&display=swap");
